Clean: Rebooting PYSAT program
The layout was not coinciding correctly with the code. Organization is
key, and the complexity is beginning to elicit bad programming. In order
to maintain readable and working code, I am going to try and reboot the
project. This will be in its own separate folder, if this is not
successful the folder will be deleted, and I’ll try working from where I
left off in V1


Former-commit-id: 2ff0072d942a2f016992f8a4815c2464f2a49edb