Continuation on saving feature for workflow
	deleted:    src/Python PYSAT_GUI/PYSAT_UI_MODULES/delete_module_.py
This was empty for some reason. Marked unnecessary
	modified:   src/Python PYSAT_GUI/PYSAT_UI_MODULES/restore_.py
line 111 to line 112
	modified:   src/Python PYSAT_GUI/pysat_ui.py
Adding in triggeres for workflow calls, open and save
adding more functionality to saveworkflow(), untested. the hope is that this method will save the current gui's state
adding more functionality to openworkflow(), untested. the hope is that this method will recall the saved state.


Former-commit-id: 69d63093b14a2de06ccbd48f5e2105c03a1011fa