Fixing CV parameters
For certain algorithms (Lasso, Elastic Net, etc), modified UI so that
the user specifies a max and min alpha and a number of intermediate
values. The program then creates a log-distributed distribution. This
mimics what is done in the internal sklearn CV, but allows the user to
see all the CV results, not just the optimum.


Former-commit-id: 6143deccb3cac5407c1dbfcbe6f72ba012ffc1d1