Nested for loop is not the right way to go
After some time figuring out how this is working, I came to the realization that we are iterating through each `_value` and then iterating through those values again. This ends up being a O(n^2) complexity.

It is better to place all the items into the widget list first, and then check the existence of the item in that list and select it. This makes the function a O(2n) complexity this will take significantly less time to compute


Former-commit-id: c73608638a907c6e96bccedba42dab78c93d0411 [formerly f0e0975339810062fae69bfdd184d9d97fbeb066]
Former-commit-id: a8bce3f18d6260729cdaa059dc6459ac8fba3197