We can now have DataBox start independently
I've also taken away the multithreading from DataBox. The reason for this has to do with the fact that Databox is now on it's own Form Thread. In other words we do not have to worry about it affecting the main GUI when we are trying to load some data, thus it is alright that we don't utilize multithreading here.


Former-commit-id: 5d42646b2c4e5f5b89b96ac2f2ec54bf8512ecb8 [formerly 9f5e578750cfbde6eea8df41919ddbd80e25f3c9]
Former-commit-id: 3661d9433f1cea8831ba4335dfcf83f7d15427f7