Added initial propagation functionality.
Currently, the only information being propagated is the "current data," which is
the most recent dataset that was used / produced.

"Current data" propagation currently works for the following modules:
Baseline Removal
Combine DataSets
Dimensionality Reduction
Load Data
Regression Train

Currently, information is propagated when the user clicks "ok," which ensures that
the data exists before propagating the request.  In other words, data is not propagated until
the user runs the program.  This should be easy to change if we want propagation on edit, and
will involve a signal on data change instead of data processing.