LOCK should prevent issues with recursion
in some really weird instances methods that change combo boxes will cause the UI to think that the UI's state is being changed, and that this warrants a run of setupModules(). This presents itself as a stackoverflow. The way to combat this is to lock down the UI temporarily until the changes to the combo boxes is done